// packages/mpl-core-kit-lib/src/generated/types/baseAppData.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import { type Codec, combineCodec, type Decoder, type Encoder, getStructDecoder, getStructEncoder } from '@solana/kit'

import {
  type BasePluginAuthority,
  type BasePluginAuthorityArgs,
  type ExternalPluginAdapterSchema,
  type ExternalPluginAdapterSchemaArgs,
  getBasePluginAuthorityDecoder,
  getBasePluginAuthorityEncoder,
  getExternalPluginAdapterSchemaDecoder,
  getExternalPluginAdapterSchemaEncoder,
} from './index.ts'

export type BaseAppData = {
  dataAuthority: BasePluginAuthority
  schema: ExternalPluginAdapterSchema
}

export type BaseAppDataArgs = {
  dataAuthority: BasePluginAuthorityArgs
  schema: ExternalPluginAdapterSchemaArgs
}

export function getBaseAppDataCodec(): Codec<BaseAppDataArgs, BaseAppData> {
  return combineCodec(getBaseAppDataEncoder(), getBaseAppDataDecoder())
}

export function getBaseAppDataDecoder(): Decoder<BaseAppData> {
  return getStructDecoder([
    ['dataAuthority', getBasePluginAuthorityDecoder()],
    ['schema', getExternalPluginAdapterSchemaDecoder()],
  ])
}

export function getBaseAppDataEncoder(): Encoder<BaseAppDataArgs> {
  return getStructEncoder([
    ['dataAuthority', getBasePluginAuthorityEncoder()],
    ['schema', getExternalPluginAdapterSchemaEncoder()],
  ])
}